Add script to build custom images and run them with host user
There are two scripts 'docker-kicad-build' to locally build image
with specific version of KiCad and Ubuntu and another
'docker-kicad-start' to start it.

The current host user is created in container and scripts will map
user's home into container. The KiCad is then run under same user
and not root which avoids problems with sharing files between
container and host system.